# cloud-brain/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zuralog Cloud Brain",
    "description": "AI Health Assistant Backend \u2014 Cross-app reasoning and autonomous actions.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analytics/daily-summary": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Daily Summary",
        "description": "Get aggregated health data for a single day.\n\nCombines activity, nutrition, sleep, and weight data into a\nunified daily snapshot.\n\nArgs:\n    user_id: The user's unique identifier.\n    date_str: Optional ISO-8601 date string. Defaults to today.\n    db: Injected async database session.\n\nReturns:\n    DailySummaryResponse with day's health aggregates.\n\nRaises:\n    HTTPException: 400 if date_str cannot be parsed.",
        "operationId": "daily_summary_api_v1_analytics_daily_summary_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          },
          {
            "name": "date_str",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 date (YYYY-MM-DD). Defaults to today.",
              "title": "Date Str"
            },
            "description": "ISO-8601 date (YYYY-MM-DD). Defaults to today."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/weekly-trends": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Weekly Trends",
        "description": "Get 7-day trend data for dashboard charts.\n\nReturns parallel arrays of daily values for the most recent\n7 days, suitable for multi-series chart rendering.\n\nArgs:\n    user_id: The user's unique identifier.\n    db: Injected async database session.\n\nReturns:\n    WeeklyTrendsResponse with 7 days of data arrays.",
        "operationId": "weekly_trends_api_v1_analytics_weekly_trends_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyTrendsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/correlation/sleep-activity": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Sleep Activity Correlation",
        "description": "Analyze correlation between sleep and activity.\n\nComputes the Pearson correlation coefficient between sleep hours\nand activity calories over the requested window, with optional\nlag offset.\n\nArgs:\n    user_id: The user's unique identifier.\n    days: Number of historical days to analyze (7-365).\n    lag: Day offset for activity relative to sleep (0-7).\n    db: Injected async database session.\n\nReturns:\n    CorrelationResponse with score, message, and data points.",
        "operationId": "sleep_activity_correlation_api_v1_analytics_correlation_sleep_activity_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "description": "Lookback days",
              "default": 30,
              "title": "Days"
            },
            "description": "Lookback days"
          },
          {
            "name": "lag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 7,
              "minimum": 0,
              "description": "Day lag for activity",
              "default": 0,
              "title": "Lag"
            },
            "description": "Day lag for activity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrelationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/trend/{metric}": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Metric Trend",
        "description": "Detect trend direction for a single health metric.\n\nCompares the recent window average against the previous window\nto classify the metric as trending up, down, or stable.\n\nArgs:\n    metric: Metric name (e.g. 'steps', 'calories_consumed').\n    user_id: The user's unique identifier.\n    window_size: Size of each comparison window (3-30).\n    db: Injected async database session.\n\nReturns:\n    TrendResponse with trend direction and statistics.",
        "operationId": "metric_trend_api_v1_analytics_trend__metric__get",
        "parameters": [
          {
            "name": "metric",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Metric"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          },
          {
            "name": "window_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 3,
              "description": "Window size",
              "default": 7,
              "title": "Window Size"
            },
            "description": "Window size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/goals": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Goals",
        "description": "Get progress for all active user goals.\n\nReturns current progress toward each of the user's active goals,\nincluding percentage completion and remaining deficit.\n\nArgs:\n    user_id: The user's unique identifier.\n    db: Injected async database session.\n\nReturns:\n    A list of GoalProgressResponse for each active goal.",
        "operationId": "get_goals_api_v1_analytics_goals_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoalProgressResponse"
                  },
                  "title": "Response Get Goals Api V1 Analytics Goals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Create Or Update Goal",
        "description": "Create or update a user goal (upsert by user_id + metric).\n\nIf a goal already exists for the given user and metric, updates\nits target value and period. Otherwise creates a new goal.\n\nArgs:\n    body: Goal creation/update request with metric, target, period.\n    user_id: The user's unique identifier.\n    db: Injected async database session.\n\nReturns:\n    GoalProgressResponse reflecting the saved goal's current progress.",
        "operationId": "create_or_update_goal_api_v1_analytics_goals_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserGoalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalProgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/dashboard-insight": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Dashboard Insight",
        "description": "Get the dashboard insight of the day.\n\nSynthesizes goal progress and metric trends into a single\nprioritized human-readable insight, along with structured\ngoal and trend data.\n\nArgs:\n    user_id: The user's unique identifier.\n    db: Injected async database session.\n\nReturns:\n    DashboardInsightResponse with insight text, goals, and trends.",
        "operationId": "dashboard_insight_api_v1_analytics_dashboard_insight_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardInsightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorrelationResponse": {
        "properties": {
          "metric_x": {
            "type": "string",
            "title": "Metric X"
          },
          "metric_y": {
            "type": "string",
            "title": "Metric Y"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "lag": {
            "type": "integer",
            "title": "Lag",
            "default": 0
          },
          "data_points": {
            "type": "integer",
            "title": "Data Points",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "metric_x",
          "metric_y",
          "score",
          "message"
        ],
        "title": "CorrelationResponse",
        "description": "Correlation analysis result between two health metrics.\n\nContains the Pearson correlation coefficient and a human-readable\ninterpretation of the relationship strength.\n\nAttributes:\n    metric_x: Name of the first metric (independent variable).\n    metric_y: Name of the second metric (dependent variable).\n    score: Pearson correlation coefficient (-1.0 to 1.0).\n    message: Human-readable description of correlation strength.\n    lag: Number of days the second metric was shifted (0 = same day).\n    data_points: Number of overlapping data points used."
      },
      "DailySummaryResponse": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "steps": {
            "type": "integer",
            "title": "Steps",
            "default": 0
          },
          "calories_consumed": {
            "type": "integer",
            "title": "Calories Consumed",
            "default": 0
          },
          "calories_burned": {
            "type": "integer",
            "title": "Calories Burned",
            "default": 0
          },
          "workouts_count": {
            "type": "integer",
            "title": "Workouts Count",
            "default": 0
          },
          "sleep_hours": {
            "type": "number",
            "title": "Sleep Hours",
            "default": 0.0
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "DailySummaryResponse",
        "description": "Aggregated health data for a single day.\n\nCombines activity, nutrition, sleep, and weight data into\na single snapshot suitable for daily summary cards.\n\nAttributes:\n    date: ISO-8601 date string (YYYY-MM-DD).\n    steps: Estimated step count for the day.\n    calories_consumed: Total caloric intake in kcal.\n    calories_burned: Total activity calories burned in kcal.\n    workouts_count: Number of recorded workouts.\n    sleep_hours: Total sleep duration in fractional hours.\n    weight_kg: Most recent weight measurement, or None if absent."
      },
      "DashboardInsightResponse": {
        "properties": {
          "insight": {
            "type": "string",
            "title": "Insight"
          },
          "goals": {
            "items": {
              "$ref": "#/components/schemas/GoalProgressResponse"
            },
            "type": "array",
            "title": "Goals",
            "default": []
          },
          "trends": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TrendResponse"
            },
            "type": "object",
            "title": "Trends",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "insight"
        ],
        "title": "DashboardInsightResponse",
        "description": "Dashboard insight of the day.\n\nCombines a prioritized textual insight with structured goal\nprogress and trend data for the frontend dashboard.\n\nAttributes:\n    insight: Human-readable insight string (1-2 sentences).\n    goals: List of goal progress snapshots.\n    trends: Mapping of metric name to trend detection result."
      },
...
      "GoalProgressResponse": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "target": {
            "type": "number",
            "title": "Target"
          },
          "current": {
            "type": "number",
            "title": "Current"
          },
          "progress_pct": {
            "type": "number",
            "title": "Progress Pct"
          },
          "is_met": {
            "type": "boolean",
            "title": "Is Met"
          },
          "remaining": {
            "type": "number",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "period",
          "target",
          "current",
          "progress_pct",
          "is_met",
          "remaining"
        ],
        "title": "GoalProgressResponse",
        "description": "Progress toward a single user-defined goal.\n\nAttributes:\n    metric: The health/fitness metric being tracked.\n    period: Goal time horizon (daily, weekly, or long_term).\n    target: The user's target value for the metric.\n    current: The current accumulated value for the period.\n    progress_pct: Percentage of target achieved.\n    is_met: Whether the goal has been met or exceeded.\n    remaining: Deficit remaining to meet the target (0 if met)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrendResponse": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "percent_change": {
            "type": "number",
            "title": "Percent Change",
            "default": 0.0
          },
          "recent_avg": {
            "type": "number",
            "title": "Recent Avg",
            "default": 0.0
          },
          "previous_avg": {
            "type": "number",
            "title": "Previous Avg",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "metric",
          "trend"
        ],
        "title": "TrendResponse",
        "description": "Trend detection result for a single metric.\n\nCompares recent values against a prior window to classify\nthe metric's direction as up, down, or stable.\n\nAttributes:\n    metric: Name of the analyzed metric.\n    trend: Direction string \u2014 one of 'up', 'down', 'stable',\n        or 'insufficient_data'.\n    percent_change: Percent change between windows.\n    recent_avg: Mean of the recent window.\n    previous_avg: Mean of the previous window."
      },
...
      "UserGoalRequest": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric",
            "description": "Metric name (e.g., 'steps', 'calories_consumed')"
          },
          "target_value": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Target Value",
            "description": "Target value for the metric"
          },
          "period": {
            "type": "string",
            "pattern": "^(daily|weekly|long_term)$",
            "title": "Period",
            "description": "Goal period"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "target_value",
          "period"
        ],
        "title": "UserGoalRequest",
        "description": "Request body for creating or updating a user goal.\n\nAttributes:\n    metric: Metric name (e.g., 'steps', 'calories_consumed').\n    target_value: Numeric target the user wants to achieve (must be > 0).\n    period: Goal time horizon \u2014 must be 'daily', 'weekly', or 'long_term'."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeeklyTrendsResponse": {
        "properties": {
          "dates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dates"
          },
          "steps": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Steps"
          },
          "calories_in": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Calories In"
          },
          "calories_out": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Calories Out"
          },
          "sleep_hours": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Sleep Hours"
          }
        },
        "type": "object",
        "required": [
          "dates",
          "steps",
          "calories_in",
          "calories_out",
          "sleep_hours"
        ],
        "title": "WeeklyTrendsResponse",
        "description": "7-day trend data for dashboard charts.\n\nProvides parallel arrays of daily values for the most recent\n7 days, intended to be plotted together on a multi-series chart.\n\nAttributes:\n    dates: List of 7 ISO-8601 date strings (oldest first).\n    steps: Daily step counts corresponding to each date.\n    calories_in: Daily caloric intake values.\n    calories_out: Daily activity calories burned.\n    sleep_hours: Daily sleep durations in fractional hours."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}